+++ include/pinocchio/unsupported.hpp
@@ -5,7 +5,7 @@
 #ifndef __pinocchio_unsupported_hpp__
 #define __pinocchio_unsupported_hpp__
 
-#include "pinocchio/deprecation.hpp"
+#include "pinocchio/deprecated.hpp"
 
 #define PINOCCHIO_UNSUPPORTED PINOCCHIO_DEPRECATED
 #define PINOCCHIO_UNSUPPORTED_MESSAGE PINOCCHIO_DEPRECATED_MESSAGE