+++ CHANGELOG.md
@@ -23,6 +23,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Removed
 
 - Remove unused headers `deprecated-macros.hpp` and `deprecated-namespaces.hpp`
+- Remove header `pinocchio/deprecation.hpp`, directly use generated `pinocchio/deprecated.hpp`
 
 ## [3.7.0] - 2025-05-21